backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.0
google-generativeai==0.3.0
PyPDF2==3.0.1
python-docx==1.1.0
